Create module for handing alerts between middleware
This solution is incomplete. What I'd like to do is be able to handle alerts between redirects, but this solution doesn't accomplish this because req.locals doesn't get handed to the middleware on a redirect. One option for this is https://stackoverflow.com/a/38355597/7400944 or I could just use the connect-flash middleware for express, but that woudl mean giving up the customization I have on theflash system