Redo attendance hooks with new understanding of old data
Old data in hooks does not correspond to if the hook is before or after
the update, so I had to use '_previousDataValues' to ask it. This had such
large implications that I just rewrote all the hooks and separated them. This
passes tests that were previously failing with the old method.