Redo event hooks for previousDataValues
Just like I misunderstood hooks for attendance records, I needed to make a
change for events. Now, if an event changes it just calculate the difference
between the two event lengths and updates members if necessary. I removed the
function abnd switched to just writing the code in the hook itself